Create POS_Count.py
I'm reusing a lot of Aaron's code for the processing.  My three functions are below the code Aaron wrote.  The functions seem really basic and that kind of saddens me because of how much time I spent researching how to do this and testing it with the large input file Aaron also created (input.txt).  There are still problems such as the tokens that are categorized as NNP (ie: ']' is an NNP for some reason), but at least these functions are able to create a count of all the POS_Tags with a POS of noun.